refactoring - relocate useEffect() which dispatch getCardsPacksTC  from PackList.tsx, Search.tsx, SortPack.tsx in Home.tsx
