+++ src/api/config/config.ts
@@ -3,6 +3,6 @@ import axios from "axios";
 // https://neko-back.herokuapp.com/2.0
 // http://localhost:7542/2.0/,
 export const instance = axios.create({
-    baseURL: "https://neko-back.herokuapp.com/2.0",
+    baseURL: "http://localhost:7542/2.0/",
     withCredentials: true,
 })
